+++ lib/src/collection/identical_items_iterable.dart
@@ -1,13 +1,12 @@
 import 'identical_items_iterator.dart';
 
 /// An efficient iterable with [length] identical elements.
-/// The value of all elements is [value].
 ///
 /// The number of elements be larger than 0.
 /// If an unsuitable [length is provided, the default value of 1
 /// is used instead.
-class SingleValueIterable<E> extends Iterable<E> {
-  const SingleValueIterable({required this.value, int length = 1})
+class IdenticalItemsIterable<E> extends Iterable<E> {
+  const IdenticalItemsIterable({required this.value, int length = 1})
     : length = (length > 0) ? length : 1;
 
   /// The value of the identical elements in this iterable.