+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.1
-- Amended docs. 
+- Amended docs.
+- Renamed `SingleValueIterable` -> `IdenticalItemsIterable`.
 
 ## 1.0.0
 - Initial version.